move token rate limit to non-ee
