+++ backend/danswer/db/connector_credential_pair.py
@@ -28,6 +28,7 @@ from danswer.utils.logger import setup_logger
 from ee.danswer.db.external_perm import delete_user__ext_group_for_cc_pair__no_commit
 from ee.danswer.external_permissions.sync_params import check_if_valid_sync_source
 
+
 logger = setup_logger()
